Moved flash-firmware to flash firmware subcommand
